sdF

using wild bootstrap to obtain standard deviation

sdF(B, formla, xformla, data, tvals, h, cl = 1)

Arguments

B
number of bootstrap iterations

formla
a formula y ~ treatment

xformla
one sided formula for x variables to include, e.g. ~x1 + x2

data
the data.frame where y, t, and x are

tvals
a grid of values of treatment variable

h
bandwidth

cl
the number of clusters to use, default is 1

Value

sd

Examples

data(igm)
igm$hs=ifelse(igm$HEDUC=="HS",1,0)
igm$col=ifelse(igm$HEDUC=="COL",1,0)
formla=lcfincome~lfincome
xformla=~hs+col
tvals=seq(quantile(igm$lfincome,probs = 0.1),quantile(igm$lfincome,probs = 0.9),length.out = 10)
h=1.2
data=igm
B=7
sdF(B,formla=formla, xformla=xformla, data=data,tvals=tvals,h=h)

#>  [1] 0.06397320 0.13758739 0.06336711 0.04779337 0.06097219 0.11789059
#>  [7] 0.06824531 0.07012007 0.06705934 0.06595768
#> attr(,"class")
#> [1] "sdF"