test.lige

test if the local intergnerational elasticity is the same across all values of the treatment variable

test.lige(cfaobj, allt, se = T, h)

Arguments

cfaobj
a CFA object

allt
all the values of the treatment variable in the dataset

se
boolean whether or not to compute standard errors

h
a bandwidth

Value

a CFASE object

Examples

# NOT RUN {
## obtain counterfactual results
out <- cfa2(lcfincome ~ lfincome, tvals, yvals, igm, method1="qr",
xformla2=~HEDUC, method2="qr", iters=100, tau1=seq(.05,.95,.05),
tau2=seq(.05,.95,.05))
test.lige(out$cfa1, allt=igm$lfincome, h=0.5)
# }